Updated functions

+++ BackEnd/controllers/recurso.js
@@ -1,8 +1,8 @@
 var Recurso = require('../models/recurso')
 
 // lista de recursos
-module.exports.list = () => {
-    return Recurso.find().sort({data: -1})
+module.exports.listRecursos = () => {
+    return Recurso.find().sort({dataCriacao: -1})
         .then(dados => {
             return dados
         })
@@ -12,7 +12,7 @@ module.exports.list = () => {
 
 }
 
-//GET /api/Recursos/:id
+//GET /api/recursos/:id
 module.exports.getRecurso = id => {
     return Recurso.findOne({_id: id})
     .then(dados => {
@@ -23,7 +23,7 @@ module.exports.getRecurso = id => {
     })
 }
 
-//POST /api/Recursos
+//POST /api/recursos
 module.exports.addRecurso = l => {
     return Recurso.collection.insertOne(l)
     .then(dados => {
@@ -34,25 +34,58 @@ module.exports.addRecurso = l => {
     })
 }
 
-//PUT /recursos/:id
-module.exports.updateRecurso = l => {
-    return Recurso.updateOne({_id: l._id}, l)
-    .then(dados => {
-        return dados
-    })
-    .catch(erro => {
-        return erro
-    })
+//PUT api/:idUser/recursos/:id
+module.exports.updateRecurso = (idUser,l) => {
+    if (l._id == idUser) {
+        return Recurso.updateOne({_id: l._id}, l)
+        .then(dados => {
+            return dados
+        })
+        .catch(erro => {
+            return erro
+        })
+    }
+    else {
+        console.log("Não é possivel alterar o recurso de outro utilizador")
+        return "Não é possivel alterar o recurso de outro utilizador"
+    }
 }
 
-//DELETE /recursos/:id
-module.exports.deleteRecurso = id => {
-    return Recurso.collection.deleteOne({_id: id})
-    .then(dados => {
-        return dados
-    })
-    .catch(erro => {
-        return erro
-    })
+//DELETE api/:idUser/recursos/:id
+module.exports.deleteRecurso = (idUser,id) => {
+    if (l._id == idUser) {
+        return Recurso.collection.deleteOne({_id: id})
+        .then(dados => {
+            return dados
+        })
+        .catch(erro => {
+            return erro
+        })
+    }
+    else{
+        console.log("Não é possivel eliminar o recurso de outro utilizador")
+        return "Não é possivel eliminar o recurso de outro utilizador"
+    }
 }
 
+//GET /api/recursos/tipos
+module.exports.tipos = () => {
+    return Recurso.distinct("recursos.tipo")
+            .then(resposta => {
+                return resposta
+            })
+            .catch(erro => {
+                return erro
+            })
+}
+
+//GET /api/recursos/tipos/:id/recursos
+module.exports.recsByTipo = (id) => {
+    return Recurso.aggregate([{$unwind: "$recursos"}, {$match: {"recursos.tipo": id}}, {$project: {"recursos.titulo":1, _id:0}}])
+            .then(resposta => {
+                return resposta
+            })
+            .catch(erro => {
+                return erro
+            })
+}
